tinyint

+++ data/sql/db-characters/base/mod_individual_spellqueue_characters.sql
@@ -2,6 +2,6 @@
 CREATE TABLE IF NOT EXISTS `mod_individual_spellqueue` (
   `guid` INT NOT NULL,
   `window` INT NOT NULL DEFAULT 400,
-  `latency` INT NOT NULL DEFAULT 0,
+  `latency` TINYINT(1) NOT NULL DEFAULT 0,
   PRIMARY KEY (`guid`)
 ) ENGINE=InnoDB DEFAULT CHARSET=UTF8MB3;
